requirements.txt
Django==3.2.11
django-ninja==0.16.1
django-cors-headers==3.11.0
django-environ==0.8.1
mysqlclient==2.1.0
pyjwt==2.3.0
requests==2.27.1
django-redis==5.2.0

# debug
ipdb
django_extensions==3.1.5
Werkzeug==2.0.2
